ZSH: Configure ZDOTDIR
Zsh config is now stored almost entirely in ~/.config/zsh/ wrt to the
XDG standard. However, the ~/.zshenv file still needs to set a zdotdir
and then source it. This is handled by the shell command added to the
end of install.conf.yaml, which:

 1. Removes ~/.zshenv
 2. Adds "ZDOTDIR=$HOME/.config/zsh" and ". $ZDOTDIR/.zshenv" to
    ~/.zshenv

Additionally, since there is no longer a need to keep and source a
separate plugins folder, all the plugins and os specific rc files have
just been moved to ~/.config/zsh