ZSH: MAJOR CHANGES -- Organize ~/.config/zsh
Now all the stuff that was cluttering zshrc and zshenv is in separate
files in the respective directory ~/.config/zsh/rc or env. Zshrc and zshenv know which dir's files to source, and will do so.

OS specific configs are now sent to a specific folder which is sourced
by zshrc and zshenv