PLANCK: Add a qwerty layer
Mac OS does this weird thing with keybinds and layouts where it responds
based on the *position* of the key and not the keycode. That means that
when switching layouts, if you hard-code a layout into your keyboard it
completely screws your bindings. And so I've added a qwerty layer which
I can enable and then just tell mac to emulate colemak.

However, this qwerty layer leaves the QOL stuff like Caps Lock ->
Backspace, shift+esc on the same key, thumb enter, etc that I have