BIB: Move to ~/.bib
This has to do with the way vimtex cmopletes citatiions. For some reason
on it doesn't detect ~/Library/texmf, which is used as texhome for
Mactex. But it *does* detect ~/texmf, which can be used by the nix
installation of latex.

Either this is an issue with mactex not setting the texhome env var, or
vimtex not searching for it, etc.

Solution I came up with is to just symlink my "global" bib files to
~/.bib, or really any dir, and then use a full link to reference them in
latex files rather than just the file name.