MAIN: Breaking - Move over zsh management to nix
In the theme of using home-manager more and more, moving over my entire
zsh config with plugins and all to home-manager is fitting. The
deprecated zsh directory is still kept in the repo until I can test the
home-manager config for a while, but after it will be removed.

Run `stow -D zsh` before `home-manager switch` to use the new config.