LAUNCHD: Remove redundant path in mail-autosync
This $PATH value was added initially because I thought the script did not
detect programs in my path. The issue was actually that I hadn't set the
correct value for $PATH in my bashrc since I was using zsh. That was
fixed when I switched to home-manager so this definition is redundant
and will probably break on other systems.