fix(nvim): use surround keybindings for sandwich
The default vim sandwich keybindings override the "s" key. This is
problematic because not only is it a default vim key but it is also used
by motion plugins like sneak or leap, which I prioritize over sandwich.
Sandwich provides a keybindings option[1] that loads vim-surround's default
keybindings which are saner.

Using lazy's init table is necessary to disable default keybindings
because the sandwich_no_default_key_mappings check is run in the
sandwich plugin file, which means the variable must be defined before
the file is loaded.

[1] https://github.com/machakann/vim-sandwich/wiki/Introduce-vim-surround-keymappings